Move common input validation rules for consumers to Common namespace
