+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.Application/Common/Validation/Consumers/SocialSecurityNumberMustBeValidRuleError.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.MarketRoles.Domain.SeedWork;
 
-namespace Energinet.DataHub.MarketRoles.Application.Consumers.Validation
+namespace Energinet.DataHub.MarketRoles.Application.Common.Validation.Consumers
 {
     public class SocialSecurityNumberMustBeValidRuleError : ValidationError
     {
