+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.Infrastructure/EDI/Errors/Converters/VATNumberMustBeValidRuleErrorConverter.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.MarketRoles.Application.Consumers.Validation;
+using Energinet.DataHub.MarketRoles.Application.Common.Validation.Consumers;
 
 namespace Energinet.DataHub.MarketRoles.Infrastructure.EDI.Errors.Converters
 {
